pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chromatic-cli"
version = "0.1.0"
description = "Python CLI/library for Chromatic dataset ingestion, indexing, and grouping."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "typer>=0.9.0",
  "pydantic>=2.0",
  "rich>=13.0",
  "sqlite-utils>=3.31",
  "requests>=2.32.0",
  "Pillow>=10.0",
  "numpy>=1.24",
  "scipy>=1.11",
  "torch>=2.0",
  "httpx>=0.25",
  "open-clip-torch>=2.24",
  "python-dotenv>=0.21",
  "pydantic-settings>=2.0",
  "platformdirs>=4.0",
  "opencv-python-headless>=4.8",
  "transformers>=4.44",
  "pypdfium2>=4.30",
  "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = ["black", "ruff"]

[project.scripts]
chromatic-cli = "chromatic_cli.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["chromatic_cli*"]
exclude = ["datasets*"]

